requirements.txt
# requirements.txt — CNN CIFAR-10 API + Évaluation

# Deep Learning
tensorflow>=2.13.0

# API Flask
flask>=3.0.0
Pillow>=10.0.0         # traitement d'images (PIL)

# Évaluation & visualisation
numpy>=1.24.0
scikit-learn>=1.3.0    # classification_report, confusion_matrix, roc_curve
matplotlib>=3.7.0
seaborn>=0.12.0
